+++ temp-04/src/core/errors/errors.handler.ts
@@ -1,8 +1,8 @@
 import { Response } from 'express';
 import { RepositoryNotFoundError } from './repository-not-found.error';
 import { HttpStatus } from '../types/http-statuses';
-import { createErrorMessages } from '../middlewares/validation/input-validtion-result.middleware';
 import { DomainError } from './domain.error';
+import { createErrorMessages } from './create-error-messages';
 
 export function errorsHandler(error: unknown, res: Response): void {
   if (error instanceof RepositoryNotFoundError) {
