+++ temp-04/src/core/errors/types/validationError.dto.ts
@@ -1,4 +1,4 @@
-import { HttpStatus } from './http-statuses';
+import { HttpStatus } from '../../types/http-statuses';
 
 type ValidationErrorOutput = {
   status: HttpStatus;
