+++ temp-04/src/core/middlewares/validation/input-validtion-result.middleware.ts
@@ -4,22 +4,9 @@ import {
   validationResult,
 } from 'express-validator';
 import { NextFunction, Request, Response } from 'express';
-import { ValidationErrorType } from '../../types/validationError';
+import { ValidationErrorType } from '../../errors/types/validationError';
 import { HttpStatus } from '../../types/http-statuses';
-import { ValidationErrorListOutput } from '../../types/validationError.dto';
-
-export const createErrorMessages = (
-  errors: ValidationErrorType[],
-): ValidationErrorListOutput => {
-  return {
-    errors: errors.map((error) => ({
-      status: error.status,
-      detail: error.detail, //error message
-      source: { pointer: error.source ?? '' }, //error field
-      code: error.code ?? null, //domain error code
-    })),
-  };
-};
+import { createErrorMessages } from '../../errors/create-error-messages';
 
 const formaValidationError = (error: ValidationError): ValidationErrorType => {
   const expressError = error as unknown as FieldValidationError;
